fix: address mcp hub review followups
