+++ backlog/tasks/task-223.2 - PR-2-MCP-Hub-setup-polish-and-diagnostics.md
@@ -22,6 +22,7 @@ modified_files:
 - backlog/tasks/task-223.2 - PR-2-MCP-Hub-setup-polish-and-diagnostics.md
 - tldw_Server_API/app/api/v1/schemas/mcp_hub_schemas.py
 - tldw_Server_API/app/api/v1/endpoints/mcp_hub_management.py
+- tldw_Server_API/app/core/MCP_unified/mcp_hub_readiness.py
 - tldw_Server_API/tests/MCP_unified/test_mcp_hub_management_api.py
 - apps/packages/ui/src/services/tldw/mcp-hub.ts
 - apps/packages/ui/src/services/tldw/__tests__/mcp-hub.test.ts
@@ -105,4 +106,6 @@ Task 7 made the MCP Hub top cards truthful by demoting the static status summary
 Task 8 added isolated MCP Hub E2E verification and final checks. The tier-2 Playwright spec now route-mocks MCP Hub readiness, external-server inventory, discovery refresh, and tool-registry responses for page load, first-run guided local stdio setup, and desktop/mobile responsive diagnostics/catalog coverage. The first-run smoke verifies empty setup state, setup choices, minimum stdio fields, Save and discover tools, mocked refresh readiness, result panel, and Tool Catalog ready state without runtime DB writes. The responsive test runs at 1440x900 and 390x844, opens guided setup choices and diagnostics, checks sanitized config redaction, switches to Tool Catalog, asserts no horizontal overflow, and attaches `mcp-hub-setup-*` / `mcp-hub-catalog-*` screenshots during Playwright. Verification: focused MCP Hub frontend suite passed with 92 tests; backend MCP Hub management API tests passed with 39 tests; `TLDW_WEB_CMD='bun run dev:webpack -- -p 8080' bun run e2e:pw -- e2e/workflows/tier-2-features/mcp-hub.spec.ts --project=tier-2 --reporter=line` passed with 5 mocked/local tests and 5 live-backend API checks skipped because no backend was available. Initial sandboxed E2E failed on localhost binding and the default Turbopack dev server failed with an internal directory-read panic, so the verified run used the existing webpack dev fallback. Full `bun run verify:design-system-state` still fails on existing repo-wide baseline findings outside this slice; MCP Hub file-scoped guard reported exitCode 0, blocked 0, baselineErrors 0. `rg "rgba\\(0,0,0,0\\.45\\)" apps/packages/ui/src/components/Option/MCPHub` returned no matches, and `git diff --check` passed. Broad Bandit wrote `/tmp/bandit_mcp_hub_setup_recovery.json` and exited 1 for existing unrelated service findings in `admin_e2e_support_service.py`, `admin_guardrails_service.py`, and `quality_eval_scheduler.py`; Task 8 itself changed only frontend E2E TypeScript.
 
 PR rebase follow-up rebased PR #2536 onto `origin/dev`, split the frontend refresh clients into bulk discovery refresh and per-server readiness refresh helpers, preserved Tool Catalog runtime query invalidation after refresh, renamed backend refresh route handlers to distinguish per-server and bulk endpoints, and tightened the MCP Hub E2E page-object Command locator after the global command-palette button made the old label query ambiguous. Verification after rebase: focused MCP Hub Vitest suite passed with 93 tests; backend MCP Hub management API tests passed with 49 tests; Bandit on `mcp_hub_management.py` exited 0 and wrote `/tmp/bandit_mcp_hub_rebase.json`; `TLDW_WEB_CMD='bun run dev:webpack -- -p 8080' bun run e2e:pw -- e2e/workflows/tier-2-features/mcp-hub.spec.ts --project=tier-2 --reporter=line` passed with 5 tests and 5 live-backend checks skipped; `git diff --check` passed.
+
+PR review follow-up addressed Qodo's latest unresolved comments after the rebase. Readiness policy and refresh-result sanitization moved from the FastAPI endpoint into `tldw_Server_API/app/core/MCP_unified/mcp_hub_readiness.py`, the validate route/UI copy now presents a read-only readiness check instead of claiming real validation, SSE auth templates use header targets, Tool Catalog empty-state logic considers visible readiness/inventory rows rather than only active totals, schema/docstring and test-marker issues were corrected, and focused regression tests were added. Verification: Python compile passed on the touched backend files with the main virtualenv; focused MCP Hub Vitest suite passed with 95 tests; backend MCP Hub management API tests passed with 49 tests; Bandit on touched backend scope exited 0 and wrote `/tmp/bandit_mcp_hub_review_followups.json`; `git diff --check` passed.
 <!-- SECTION:FINAL_SUMMARY:END -->
