+++ tldw_Server_API/app/api/v1/endpoints/mcp_hub_management.py
@@ -69,7 +69,6 @@ from tldw_Server_API.app.api.v1.schemas.mcp_hub_schemas import (
     GovernanceAuditFindingListResponse,
     MCPHubDeleteResponse,
     McpCredentialSlotStatusResponse,
-    McpHubDiscoveryRefreshResultResponse,
     McpHubReadinessResponse,
     McpServerReadinessResponse,
     PathScopeObjectCreateRequest,
@@ -104,6 +103,12 @@ from tldw_Server_API.app.core.AuthNZ.principal_model import AuthPrincipal
 from tldw_Server_API.app.core.AuthNZ.repos.mcp_hub_repo import McpHubRepo
 from tldw_Server_API.app.core.config import config
 from tldw_Server_API.app.core.exceptions import BadRequestError, ResourceNotFoundError
+from tldw_Server_API.app.core.MCP_unified.mcp_hub_readiness import (
+    build_hub_readiness_payload,
+    build_server_readiness_payload,
+    is_available_for_discovery_refresh,
+    sanitize_discovery_refresh_result_payload,
+)
 from tldw_Server_API.app.core.MCP_unified.server import get_mcp_server
 from tldw_Server_API.app.services.mcp_hub_external_legacy_inventory import (
     McpHubExternalLegacyInventoryService,
@@ -171,44 +176,6 @@ _WINDOWS_ABSOLUTE_PATH_RE = re.compile(r"^[A-Za-z]:[\\/]")
 _PATH_SCOPE_DOCUMENT_KEYS = {"path_scope_mode", "path_scope_enforcement", "path_allowlist_prefixes"}
 _MCP_HUB_REFRESH_TOOL_NAME = "external.tools.refresh"
 _MCP_HUB_REFRESH_LOCK_TIMEOUT_SECONDS = 30.0
-_MCP_HUB_READINESS_REASON_PRIORITY = [
-    "auth_missing",
-    "runtime_unavailable",
-    "preflight_failed",
-    "unreachable",
-    "discovery_failed",
-    "config_changed",
-    "discovery_not_run",
-    "no_tools_returned",
-    "catalog_expired",
-    "partial_capability",
-]
-_MCP_HUB_DISPLAY_STATE_BY_REASON = {
-    "not_configured": "needs_setup",
-    "preflight_failed": "needs_attention",
-    "discovery_not_run": "needs_attention",
-    "auth_missing": "needs_attention",
-    "runtime_unavailable": "needs_attention",
-    "unreachable": "needs_attention",
-    "discovery_failed": "needs_attention",
-    "no_tools_returned": "no_tools",
-    "config_changed": "stale",
-    "catalog_expired": "stale",
-    "partial_capability": "ready",
-}
-_MCP_HUB_ACTIONS_BY_REASON = {
-    "not_configured": ["add_server"],
-    "preflight_failed": ["edit_config", "validate", "view_details"],
-    "discovery_not_run": ["refresh_discovery", "edit_config"],
-    "auth_missing": ["open_credentials", "view_details"],
-    "runtime_unavailable": ["edit_config", "view_details"],
-    "unreachable": ["edit_config", "refresh_discovery", "view_details"],
-    "discovery_failed": ["refresh_discovery", "view_details"],
-    "no_tools_returned": ["refresh_discovery", "view_details"],
-    "config_changed": ["refresh_discovery", "edit_config"],
-    "catalog_expired": ["refresh_discovery", "view_details"],
-    "partial_capability": ["open_tool_catalog", "view_details"],
-}
 _mcp_hub_refresh_locks: dict[str, asyncio.Lock] = {}
 _mcp_hub_refresh_locks_guard = asyncio.Lock()
 
@@ -1153,257 +1120,6 @@ def _external_row_to_response(row: dict[str, Any]) -> ExternalServerResponse:
     )
 
 
-def _readiness_unique_reasons(reasons: list[str]) -> list[str]:
-    """Return reason codes ordered by MCP Hub readiness priority."""
-    seen = set()
-    ordered: list[str] = []
-    priority = {reason: index for index, reason in enumerate(_MCP_HUB_READINESS_REASON_PRIORITY)}
-    for reason in sorted(reasons, key=lambda value: priority.get(value, len(priority))):
-        if reason not in seen:
-            seen.add(reason)
-            ordered.append(reason)
-    return ordered
-
-
-def _readiness_primary_reason(reasons: list[str]) -> str | None:
-    """Return the highest-priority readiness reason, if any."""
-    ordered = _readiness_unique_reasons(reasons)
-    return ordered[0] if ordered else None
-
-
-def _readiness_allowed_actions(reasons: list[str]) -> list[str]:
-    """Union readiness actions for the supplied reason codes."""
-    actions: list[str] = []
-    for reason in reasons:
-        for action in _MCP_HUB_ACTIONS_BY_REASON.get(reason, []):
-            if action not in actions:
-                actions.append(action)
-    return actions
-
-
-def _readiness_message(primary_reason: str | None, credential_state: str) -> str:
-    """Return a user-facing readiness status message without exposing config details."""
-    if primary_reason == "auth_missing":
-        return "Credentials are required before this server can be used."
-    if primary_reason == "runtime_unavailable":
-        return "Runtime is not available for this server."
-    if primary_reason == "preflight_failed":
-        return "Preflight validation failed. Check the server configuration."
-    if primary_reason == "unreachable":
-        return "Server cannot be reached."
-    if primary_reason == "discovery_failed":
-        return "Discovery ran but failed."
-    if primary_reason == "config_changed":
-        return "Server config or discovery state changed. Refresh discovery."
-    if primary_reason == "discovery_not_run":
-        if credential_state == "not_required":
-            return "No credentials required. Discover tools to make this server available."
-        return "Server is saved, but tool discovery has not run."
-    if primary_reason == "no_tools_returned":
-        return "Server responded, but exposed no tools."
-    if primary_reason == "catalog_expired":
-        return "Tool catalog is stale. Refresh discovery."
-    if primary_reason == "partial_capability":
-        return "Ready with limited capability."
-    if primary_reason == "not_configured":
-        return "Add an external server to start MCP Hub setup."
-    return "Ready. No credentials required." if credential_state == "not_required" else "Ready."
-
-
-def _credential_state_for_external_row(row: dict[str, Any]) -> str:
-    """Normalize external-server credential state without returning secret material."""
-    slots = [slot for slot in (row.get("credential_slots") or []) if isinstance(slot, dict)]
-    required_slots = [slot for slot in slots if bool(slot.get("is_required", True))]
-    if any(not bool(slot.get("secret_configured")) for slot in required_slots):
-        return "required_missing"
-    if any(bool(slot.get("secret_configured")) for slot in slots) or row.get("auth_template_valid") is True:
-        return "configured"
-    has_auth_template = bool(row.get("auth_template_present"))
-    if bool(row.get("secret_configured")) and not has_auth_template and not slots:
-        return "legacy_fallback"
-    blocked_reason = str(row.get("auth_template_blocked_reason") or "").strip()
-    if (
-        str(row.get("transport") or "").strip().lower() == "stdio"
-        and not has_auth_template
-        and not slots
-        and blocked_reason in {"", "no_auth_template"}
-    ):
-        return "not_required"
-    return "unknown"
-
-
-def _is_authoritative_external_tool(entry: dict[str, Any], server_id: str) -> bool:
-    """Return True only for registry signals that authoritatively belong to an external server."""
-    tool_name = str(entry.get("tool_name") or "")
-    module_name = str(entry.get("module") or "")
-    return tool_name.startswith(f"ext.{server_id}.") or module_name == f"external.{server_id}"
-
-
-def _matching_external_tool_entries(
-    registry_entries: list[dict[str, Any]],
-    *,
-    server_id: str,
-) -> list[dict[str, Any]]:
-    return [entry for entry in registry_entries if _is_authoritative_external_tool(entry, server_id)]
-
-
-def _operation_metadata_from_row(row: dict[str, Any]) -> dict[str, Any] | None:
-    current_operation = row.get("current_operation")
-    return current_operation if isinstance(current_operation, dict) else None
-
-
-def _coerce_nonnegative_int(value: Any) -> int:
-    try:
-        return max(0, int(value))
-    except (TypeError, ValueError):
-        return 0
-
-
-def _extract_refresh_payload(raw_result: dict[str, Any]) -> dict[str, Any]:
-    """Extract the manager refresh payload from direct or MCP content-wrapped results."""
-    if not isinstance(raw_result, dict):
-        return {}
-    if any(key in raw_result for key in ("refreshed_servers", "total_servers", "virtual_tools", "errors")):
-        return raw_result
-    result = raw_result.get("result")
-    if isinstance(result, dict):
-        extracted = _extract_refresh_payload(result)
-        if extracted:
-            return extracted
-    content = raw_result.get("content")
-    if isinstance(content, list):
-        for item in content:
-            if not isinstance(item, dict):
-                continue
-            json_payload = item.get("json")
-            if isinstance(json_payload, dict):
-                return json_payload
-    return {}
-
-
-def _sanitize_refresh_result(raw_result: dict[str, Any]) -> McpHubDiscoveryRefreshResultResponse:
-    """Return refresh metadata with generic error messages only."""
-    payload = _extract_refresh_payload(raw_result)
-    raw_errors = payload.get("errors") if isinstance(payload.get("errors"), dict) else {}
-    errors = {str(server_id): "Discovery refresh failed." for server_id in raw_errors}
-    return McpHubDiscoveryRefreshResultResponse(
-        refreshed_servers=_coerce_nonnegative_int(payload.get("refreshed_servers")),
-        total_servers=_coerce_nonnegative_int(payload.get("total_servers")),
-        virtual_tools=_coerce_nonnegative_int(payload.get("virtual_tools")),
-        errors=errors,
-    )
-
-
-def _server_readiness_response(
-    row: dict[str, Any],
-    *,
-    registry_entries: list[dict[str, Any]],
-    current_operation: dict[str, Any] | None = None,
-    last_validation_at: datetime | str | None = None,
-    refresh_result: McpHubDiscoveryRefreshResultResponse | None = None,
-) -> McpServerReadinessResponse:
-    """Build sanitized readiness response for one external server row."""
-    server_id = str(row.get("id") or "")
-    server_name = str(row.get("name") or server_id)
-    credential_state = _credential_state_for_external_row(row)
-    matching_entries = _matching_external_tool_entries(registry_entries, server_id=server_id)
-    tool_count = len(matching_entries)
-    operation = current_operation or _operation_metadata_from_row(row)
-    if operation:
-        operation_type = str(operation.get("operation_type") or "validation")
-        if operation_type not in {"validation", "discovery"}:
-            operation_type = "validation"
-        return McpServerReadinessResponse(
-            server_id=server_id,
-            server_name=server_name,
-            display_state="checking",
-            credential_state=credential_state,
-            tool_count=tool_count,
-            reason_codes=[],
-            primary_reason_code=None,
-            allowed_actions=["view_details"],
-            message=(
-                "Tool discovery is running."
-                if operation_type == "discovery"
-                else "Preflight validation is running."
-            ),
-            current_operation={
-                "operation_type": operation_type,
-                "started_at": operation.get("started_at"),
-                "message": operation.get("message"),
-            },
-            last_validation_at=last_validation_at or row.get("last_validation_at"),
-            last_discovery_at=row.get("last_discovery_at"),
-            last_successful_discovery_at=row.get("last_successful_discovery_at"),
-        )
-
-    reasons: list[str] = []
-    if credential_state == "required_missing":
-        reasons.append("auth_missing")
-    if not bool(row.get("enabled")) or row.get("runtime_executable") is False:
-        reasons.append("runtime_unavailable")
-
-    blocked_reason = row.get("auth_template_blocked_reason")
-    if (row.get("auth_template_present") and row.get("auth_template_valid") is False) or (
-        blocked_reason and str(blocked_reason) != "no_auth_template"
-    ):
-        reasons.append("preflight_failed")
-
-    last_error_category = row.get("last_error_category")
-    last_error_message = row.get("last_error_message")
-    if refresh_result is not None and server_id in refresh_result.errors:
-        reasons.append("discovery_failed")
-        last_error_category = "discovery_failed"
-        last_error_message = "Discovery refresh failed."
-    elif last_error_category == "discovery_failed":
-        reasons.append("discovery_failed")
-        last_error_message = "Discovery refresh failed."
-
-    if tool_count == 0 and "runtime_unavailable" not in reasons and "discovery_failed" not in reasons:
-        reasons.append("discovery_not_run")
-    if tool_count > 0 and any(entry.get("metadata_warnings") for entry in matching_entries):
-        reasons.append("partial_capability")
-
-    reason_codes = _readiness_unique_reasons(reasons)
-    primary_reason = _readiness_primary_reason(reason_codes)
-    display_state = _MCP_HUB_DISPLAY_STATE_BY_REASON.get(primary_reason or "", "ready")
-    allowed_actions = (
-        _readiness_allowed_actions(reason_codes)
-        if reason_codes
-        else ["open_tool_catalog", "view_details"]
-    )
-    return McpServerReadinessResponse(
-        server_id=server_id,
-        server_name=server_name,
-        display_state=display_state,
-        credential_state=credential_state,
-        tool_count=tool_count,
-        reason_codes=reason_codes,
-        primary_reason_code=primary_reason,
-        allowed_actions=allowed_actions,
-        message=_readiness_message(primary_reason, credential_state),
-        current_operation=None,
-        last_validation_at=last_validation_at or row.get("last_validation_at"),
-        last_discovery_at=row.get("last_discovery_at"),
-        last_successful_discovery_at=row.get("last_successful_discovery_at"),
-        last_error_category=last_error_category,
-        last_error_message=last_error_message,
-        refresh_result=refresh_result,
-    )
-
-
-def _is_operational_managed_external_row(row: dict[str, Any]) -> bool:
-    return (
-        str(row.get("server_source") or "managed") == "managed"
-        and not row.get("superseded_by_server_id")
-        and bool(row.get("enabled"))
-    )
-
-
-def _is_available_for_discovery_refresh(row: dict[str, Any]) -> bool:
-    return _is_operational_managed_external_row(row) and row.get("runtime_executable") is True
-
-
 async def _list_visible_external_server_rows(
     *,
     principal: AuthPrincipal,
@@ -1894,68 +1610,8 @@ async def get_mcp_hub_readiness(
         owner_scope_id=owner_scope_id,
     )
     registry_entries = await registry.list_entries()
-    server_readiness = [
-        _server_readiness_response(row, registry_entries=registry_entries)
-        for row in rows
-    ]
-    operational_server_ids = {
-        str(row.get("id") or "")
-        for row in rows
-        if _is_operational_managed_external_row(row)
-    }
-    operational_readiness = [
-        readiness
-        for readiness in server_readiness
-        if readiness.server_id in operational_server_ids
-    ]
-
-    if not operational_readiness:
-        return McpHubReadinessResponse(
-            display_state="needs_setup",
-            reason_codes=["not_configured"],
-            primary_reason_code="not_configured",
-            allowed_actions=["add_server"],
-            message=_readiness_message("not_configured", "not_required"),
-            servers=server_readiness,
-            total_servers=0,
-        )
-
-    aggregate_reasons = _readiness_unique_reasons(
-        [
-            reason
-            for readiness in operational_readiness
-            for reason in readiness.reason_codes
-        ]
-    )
-    primary_reason = _readiness_primary_reason(aggregate_reasons)
-    if any(readiness.display_state == "checking" for readiness in operational_readiness):
-        display_state = "checking"
-    elif primary_reason:
-        display_state = _MCP_HUB_DISPLAY_STATE_BY_REASON.get(primary_reason, "needs_attention")
-    else:
-        display_state = "ready"
-
-    return McpHubReadinessResponse(
-        display_state=display_state,
-        reason_codes=aggregate_reasons,
-        primary_reason_code=primary_reason,
-        allowed_actions=(
-            _readiness_allowed_actions(aggregate_reasons)
-            if aggregate_reasons
-            else ["open_tool_catalog", "view_details"]
-        ),
-        message=(
-            _readiness_message(primary_reason, "configured")
-            if primary_reason
-            else "MCP Hub is ready."
-        ),
-        servers=server_readiness,
-        total_servers=len(operational_readiness),
-        ready_server_count=sum(1 for item in operational_readiness if item.display_state == "ready"),
-        checking_server_count=sum(1 for item in operational_readiness if item.display_state == "checking"),
-        attention_server_count=sum(1 for item in operational_readiness if item.display_state == "needs_attention"),
-        no_tool_server_count=sum(1 for item in operational_readiness if item.display_state == "no_tools"),
-        stale_server_count=sum(1 for item in operational_readiness if item.display_state == "stale"),
+    return McpHubReadinessResponse.model_validate(
+        build_hub_readiness_payload(rows, registry_entries)
     )
 
 
@@ -1969,7 +1625,7 @@ async def validate_external_server(
     svc: McpHubService = Depends(get_mcp_hub_service),
     registry: McpHubToolRegistryService = Depends(get_mcp_hub_tool_registry_dep),
 ) -> McpServerReadinessResponse:
-    """Return current server readiness after permission and visibility checks."""
+    """Return the current server readiness snapshot after visibility checks."""
     _require_mutation_permission(principal)
     row = await _get_visible_external_server_row_or_404(
         server_id=server_id,
@@ -1977,10 +1633,8 @@ async def validate_external_server(
         svc=svc,
     )
     registry_entries = await registry.list_entries()
-    return _server_readiness_response(
-        row,
-        registry_entries=registry_entries,
-        last_validation_at=datetime.now(timezone.utc),
+    return McpServerReadinessResponse.model_validate(
+        build_server_readiness_payload(row, registry_entries=registry_entries)
     )
 
 
@@ -2004,7 +1658,7 @@ async def refresh_single_external_server_discovery(
         principal=principal,
         svc=svc,
     )
-    if not _is_available_for_discovery_refresh(row):
+    if not is_available_for_discovery_refresh(row):
         raise HTTPException(status_code=409, detail="External server is not available for discovery refresh")
     lock = await _get_refresh_lock(server_id)
     try:
@@ -2016,17 +1670,19 @@ async def refresh_single_external_server_discovery(
     finally:
         lock.release()
 
-    refresh_result = _sanitize_refresh_result(raw_result)
+    refresh_result = sanitize_discovery_refresh_result_payload(raw_result)
     row = await _get_visible_external_server_row_or_404(
         server_id=server_id,
         principal=principal,
         svc=svc,
     )
     registry_entries = await registry.list_entries()
-    return _server_readiness_response(
-        row,
-        registry_entries=registry_entries,
-        refresh_result=refresh_result,
+    return McpServerReadinessResponse.model_validate(
+        build_server_readiness_payload(
+            row,
+            registry_entries=registry_entries,
+            refresh_result=refresh_result,
+        )
     )
